+++ src/maite/interop/provider_registry.py
@@ -1,0 +1,17 @@
+# Copyright 2024, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
+# Subject to FAR 52.227-11 – Patent Rights – Ownership by the Contractor (May 2014).
+# SPDX-License-Identifier: MIT
+
+from maite._internals.interop.provider_registry import (
+    create_provider,
+    get_provider_type,
+    list_providers,
+    register_provider,
+)
+
+__all__ = [
+    "list_providers",
+    "register_provider",
+    "create_provider",
+    "get_provider_type",
+]