+++ src/maite/protocols/providers.py
@@ -1,0 +1,15 @@
+# Copyright 2024, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
+# Subject to FAR 52.227-11 – Patent Rights – Ownership by the Contractor (May 2014).
+# SPDX-License-Identifier: MIT
+
+from maite._internals.protocols.providers import (
+    DatasetProvider,
+    MetricProvider,
+    ModelProvider,
+)
+
+__all__ = [
+    "ModelProvider",
+    "DatasetProvider",
+    "MetricProvider",
+]