+++ src/maite/testing/docs.py
@@ -1,4 +1,4 @@
-# Copyright 2023, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
+# Copyright 2024, MASSACHUSETTS INSTITUTE OF TECHNOLOGY
 # Subject to FAR 52.227-11 – Patent Rights – Ownership by the Contractor (May 2014).
 # SPDX-License-Identifier: MIT
 
@@ -9,8 +9,7 @@ Utilities for validating documentation strings for maite projects.
 from maite._internals.testing.docs import (
     NumpyDocErrorCode,
     NumPyDocResults,
-    person,
     validate_docstring,
 )
 
-__all__ = ["validate_docstring", "NumpyDocErrorCode", "NumPyDocResults", "person"]
+__all__ = ["validate_docstring", "NumpyDocErrorCode", "NumPyDocResults"]